Bug 1855758 - Implement better progress notifications from the parser to the front end. r=BenC,mkmelin
So that we can trigger actions as soon as the message body is ready, we need to know when the
parser has finished outputting it.

Unfortunately my original plan of using web progress notifications is inadequate as it's not
possible to pass useful information with the notifications, so this patch implements a specialist
notification. (Thanks to nsBrowserStatusFilter for this, it discards almost all the data it
receives.) The "mark message as read" action is then connected to the notification.

Differential Revision: https://phabricator.services.mozilla.com/D191039

--HG--
extra : moz-landing-system : lando