Bug 522768 - Add tooltip displaying full path in location column. r=aleca,darktrojan
Implement displaying the full path (including all parents) of a folder as a tooltip in the
location column of the thread pane.

This is achieved by providing a new readonly attribute `prettyPath` in `nsIMsgFolder` which is
returned by `nsMsgDBView::CellTextForColumn` instead of `prettyName`.

Differential Revision: https://phabricator.services.mozilla.com/D188694

--HG--
extra : amend_source : 2b124cd104292cac3a24e80407585781a55d2753