Bug 1862232 - Fix a broken test. r=mkmelin
`nsIMsgDBView.URIForFirstSelectedMessage` throws an exception when there is no selected message.
Although `numSelectedMessages` is 1, we're looking at a message loaded from a file so there are
no messages to select. (It's always 1 when we're in a message tab or message window.)

Differential Revision: https://phabricator.services.mozilla.com/D192543

--HG--
extra : amend_source : 221658f79bcdc48abc0ec6b6fdbf8c26dcee51a6