Bug 1862822 - Don't mark messages in background tabs as read. r=mkmelin
This changes uses the page visibility API to avoid marking messages as read until the tab is
switched to. First I had to mark background browsers as inactive for the API to work.

In order to test the behaviour with an IMAP server, I had to modify the IMAP fakeserver to send
message flags in the response to a FETCH command, as real servers do. Then I discovered the \Seen
flag isn't propagated if it's the only flag, which seems like a real problem that should be fixed.

Differential Revision: https://phabricator.services.mozilla.com/D192654

--HG--
extra : amend_source : a20f13257b4582dd1899d3f76d00c75af6bcd7b8