Bug 1865529 - Simplify EnigmailCore to only handle initialisation of Enigmail components. r=mkmelin
This simpifies the code down to remove the unnecessary wrapping and resolve it into a simple init function.
The calling points of Enigmail.getService are simply replaced with a call to the new init. They could possibly be reviewed
at a later stage to see if they are still necessary or not.

Differential Revision: https://phabricator.services.mozilla.com/D194008

--HG--
extra : moz-landing-system : lando