Bug 1846127 - Fix applying columns or view to a root folder and its descendants. r=mkmelin
The root folder has no message database, so it would throw an exception and the rest of the operation would fail.

Setting the column states worked in 102 because that used `folder.setStringProperty` which avoided the error.
Now we write directly to the database, which also doesn't write to the folder cache. We never used the value from the cache anyway.

Differential Revision: https://phabricator.services.mozilla.com/D194029

--HG--
extra : moz-landing-system : lando