Bug 1867003 - Port bug 1564553: Remove CMakeLists.txt from GENERATED_FILES. r=dandarnell
Bug 1564553 made some changes to how the generated files tarballs are built. It
has the effect of trying to include a CMakeLists.txt file which is generated as
part of building Botan. The file is present, just not in the directory where
the generated files packager is looking for it.

Since the file is not needed it can be safely removed from the GENERATED_FILES
list so the build system will not try to package it. Additionally, the file
is now removed from $OBJDIR right after it's created to avoid any confusion.

Bug 1621785 is CMakeLists.txt's file's origin. In order to get a usable build.h
file for Botan to compile, Botan's configure.py script is run with '--with-cmake'
to avoid overwriting the Makefile generated by mozbuild.

Differential Revision: https://phabricator.services.mozilla.com/D194892

--HG--
extra : moz-landing-system : lando