Bug 1874067 - Improve usage telemetry identifiers for anonymous elements. r=john.bieling
This creates a new convention of adding a 'data-telemetry-id' attribute to otherwise anonymous or
dynamically generated content where using the ID is not feasible or inappropriate.

It also corrects the telemetry recording of add-on browserAction and composeAction buttons (which
should be obfuscated) and tests that these are handled correctly.

Differential Revision: https://phabricator.services.mozilla.com/D198356

--HG--
extra : rebase_source : cc28ce2869779dd069326a2ceb569cbe133cb2a2