Bug 1869860 - Set up "cron" decision task to run tb-rust automation. r=dandarnell
A bit of a hack to generate a Taskcluster hook that will trigger the tb-rust
vendor automation job. The cron task will never actually run with an empty
schedule, but with appropriate grants in ci-configuration it will be triggerable
by a CI job on mozilla-central and by Thunderbird sheriffs & releng.

Differential Revision: https://phabricator.services.mozilla.com/D199632

--HG--
extra : moz-landing-system : lando