Bug 1427546 - Fix expanding threads with invisible children only. r=mkmelin
When all children of a thread are ignored subthreads, or, in unread only view, an unread top thread
message has read children only, just the message summary is shown. Neither trying to expand the
thread, nor clicking on a message in the summary displays the top message itself. This fixes both.

Differential Revision: https://phabricator.services.mozilla.com/D202119

--HG--
extra : moz-landing-system : lando