Bug 1880655 - Add ComposeDetails.isModified and allow to clear the modification flag. r=mkmelin
It is a commom practice for add-ons to prefill the composer with default
content or settings. These settings may be considdered default and the
composer should not ask for confirmation on close, if the user did not
apply any additional changes.

This patch exposes the modification status of the composer and allows to
reset it.

Differential Revision: https://phabricator.services.mozilla.com/D202729

--HG--
extra : rebase_source : 7b6e072bc451ac59cd2b4ea809c7479d10d8bae6