Bug 1854684 - Introduce the MsgHdrProcessor class. r=mkmelin
The `MsgHdrProcessor` class will be the new main entry point to get
raw or parsed data from a message. It does not depend on
`MsgHdrToMimeMessage`, but on the `MimeTreeEmitter` and the
`MimeTreeDecryptor`.

It re-implements access to the raw message, the parsed message, and
attachment retrieval.

The function `MsgHdrProcessor.getOriginalMessage()` is a clone of
`getRawMessage()`, to keep review of the follow-up patches simple (where
it gets removed).

Differential Revision: https://phabricator.services.mozilla.com/D202595

--HG--
extra : moz-landing-system : lando