Bug 1860994 - Explicitly allow "msip_labels" to be set as a custom header in setComposeDetails(). r=mkmelin
If an add-on needs to set this header, I think it should be possible.
The only red line is modifiying real message headers, which are
controled by Thunderbird already. These headers should always be
modified through the appropriate APIs (ComposeDetails.returnReceipt or
ComposeDetails.deliveryStatusNotification).

If we get requests for a lot more headers which qualify to be
accessible, we might consider inverting the principle in the future and
keep a list of headers we do NOT want to be modified through the API.

This patch also adds protection for X-Mozilla-* headers from being
manipulated through the WebExtension API.

Differential Revision: https://phabricator.services.mozilla.com/D202815

--HG--
extra : amend_source : b3de693809dc9bda10f2dc2ec1afb33e757d5261