Bug 1875633 - Avoid timeouts when huge number of pop3 messages left on server. r=mkmelin
This now sends a pop3 NOOP when processing LIST, UIDL, TOP and RETR responses if
the time since the last pop3 command sent exceeds 10 seconds. This acts as a
"keep-alive" so the server doesn't close the connection due to inactivity.

A change is also made to LineReader.jsm (also used by NNTP) so that the whole
response is received before processing the individual lines. This ensures that
NOOP is not sent while server responses may still be coming in.

Finally, this includes a fix to c++ pop3 code made at Bug 1577548 for "temporary"
errors reported by the server. The fix was never considered or included in the JS
pop3 version. Specifically, if a UIDL response error occurs and the server has UIDL
capability, UIDL command  will now be tried again on the next check for new mail
and reported as a temporary error rather than stopping further checks with a
modal dialog. (This doesn't handle possible temporary errors from other
commands, but UIDL is historically the  main cause.)

Differential Revision: https://phabricator.services.mozilla.com/D203697

--HG--
extra : amend_source : ec72235e2d93e223ff5c5129303c6ac971094889