Bug 1668754 - Force popups to use add-on icon, for better discoverability of add-on popups vs Thunderbird popups. r=aleca,mkmelin
This is a possible implementation for the icon of popups. These are only
supported by Windows.

On Windows, all popups currently use the Thunderbird logo icon. Since
popups do not have a URL bar, it is not discoverable which popup is
opened by which add-on (they usually open moz-extension:// urls, which
point to pages within the add-on).

For the menus API we have the rule, that the icon used in context menus
is always the add-on icon, so that the user is able to identify which
menu entries are created by which add-on. We could follow the same
principle and force popups to use the add-on icon.

To see the patch in action, you can install one of the recommended
add-ons, for example the "Mail Merge Add-on" or the "Auto Adress
Cleaner" from the list of recommended add-ons and then click on the gear
icon inside AMO and select "Debug Add-ons". On the new page click the
"Inspect" button of the installed add-on and switch to the console tab
and enter:

`browser.windows.create({url:"https://example.com", type:"popup"})`

This patch is not allowing to use arbitrary icons, but enforces the one
defined in the manifest of the extension. It uses the Thunderbird logo
icon, if the add-on did not specify an icon in its manifest.

Differential Revision: https://phabricator.services.mozilla.com/D204270

--HG--
extra : moz-landing-system : lando