Bug 1877390 - Rework sortController.convertSortTypeToColumnID() to sortControler.getCurrentSortColumnId(). r=freaktechnik
The function convertSortTypeToColumnID() uses dbView.curCustomColumn to
convert the "byCustom" sort type to a columnID. This however only works
for the current custom sort column, and not for any custom column. It is
also only used with the primarySortType (and not with any sort type) as
input.

To improve readability of this code, this patch reworks this function to
always return to columnID of the current sort column.

Differential Revision: https://phabricator.services.mozilla.com/D200212

--HG--
extra : rebase_source : 219d43da8faa064fdd95bc14388c1bc185aceb23