Bug 1679711 - Attempt OAuth2 against known providers with Exchange Autodiscover. r=ikey,leftmostcat
This patch does not implement Exchange's HTTP 401 Challenge described in bug 1679711, as more research is needed to figure out exactly how this can be done. Instead, it tries authenticating against known providers the "normal" way when possible, then using the resulting token against the Autodiscover endpoint. If it can't, then it tries falling back to Basic auth.

This patch does not support falling back to Basic auth if the Autodiscover endpoint responds with a 401 following an OAuth-authenticated request. This may be implemented in the future, but in the meantime being able to get an OAuth2 token but not being allowed to use it against the autodiscover endpoint should be unlikely enough that this is not an issue.

Differential Revision: https://phabricator.services.mozilla.com/D205762