Bug 1890453 - Add C++ shortcut URI accessor to nsIMsgFolder (and nsMsgDBFolder). r=mkmelin
There's a generic nsIMsgFolder::URI() which is implemented in terms of GetURI().
This will work on non nsMsgDBFolder-based folders, even if they are written in JS or rust or whatever.
Then there's also an inline nsMsgDBFolder::URI() which the compiler can optimise for when the concrete type is availible to it.

Differential Revision: https://phabricator.services.mozilla.com/D206972

--HG--
extra : amend_source : b64406dfb6e624ae28b93587e3340b43bc976dd7