Bug 1878944 - Fixing the sometimes not sent e-mail about the acceptance/rejection of appointments r=mkmelin
Currently, the email address in the account and the one used in the invitation must match one to one. This means that emails cannot be sent if there is a difference in capitalization.
This is now fixed by this code change.

Differential Revision: https://phabricator.services.mozilla.com/D207474

--HG--
extra : amend_source : 5892eae5c102b4f18876e917d7a45611797451fd