Bug 1889653 - Register a subfolder in memory before creating the storage for it. r=ikey,leftmostcat
For some folder types, nsIMsgFolder::GetSubFolders calls nsIMsgPluggableStore::DiscoverSubFolders, which might attempt to rectify its in-memory representation of the folder architecture if it finds storage for folders it hasn't already recorded. nsIMsgFolder::GetSubFolders is called by nsIMsgFolder::AddSubfolder (via nsIMsgFolder::GetChildWithURI), meaning that if, when creating the subfolder, we create the storage for it before registering it, we might end up in a situation where nsIMsgFolder::AddSubfolder tries to register it twice, and fails on the second attempt (because the subfolder already exists).

Differential Revision: https://phabricator.services.mozilla.com/D207320

--HG--
extra : amend_source : 038a9b85529ee6b15f691e682d96f4ffcb81ed38