Bug 1854779 - Support discovery of iCloud calendars. r=darktrojan
I tend to officially sunset DAV-4-TbSync, and the only remaining
"feature" it supports which is not yet supported by Thunderbird is the
detection of iCloud calendars and address books.

I poked around the source and found that the `detectCollection()`
routine already has the calendar-home-set, but it is ignored and instead
it calls `handlePrincipal()`, trying to get the home-set from another
PROPFIND request, which fails.

This patch adds an early exit and uses the already found home-set,
instead of issuing another server request.

If this is not the right fix, it may at least give an idea of what is
going on. I can provide iCloud access credentials.

With this patch, I can enter my Apple-ID as username using a dedicated
app password created for Thunderbird, and the server `icloud.com`, to
get my calendars discovered and added.

Differential Revision: https://phabricator.services.mozilla.com/D206253

--HG--
extra : amend_source : 36c57c8d8ae2a463974e54298ff494fdf3645b29