Bug 1894905 - Deprecate flat contact properties in Manifest V3 and only work with the vCard. r=#thunderbird-reviewers
This looks larger than it really is. For contacts, we currently return a
property bag, which includes flat properties like "PrimaryEmail" or
"FirstName", which is how the old address book stored contact data. We
kept supporting these properties even after the internal storage was
moved to vCard. The vCard was just another property in said property bag
and we applied some voodoo to allow extension developers to continue
to set the flat properties, or the vCard.

In MV3 we stop doing that and just work with the vCard. This affects
the functions `contacts.create()` and `contacts.update()`, and also the
events `contacts.onCreated` and `contacts.onUpdated`.

This patch is not changing MV2 behavior.

Since we no longer expose the internal property bag, this also removes
access to things like `preferDisplayName`, which is still a property of
the internal card. I believe it is bad practice to expose these
internals. If there is a request to be able to access such values,
they should be moved as X-Properties to the vCard (which would sync
these settings with the vCard between Thunderbird instances).

Differential Revision: https://phabricator.services.mozilla.com/D209384

--HG--
extra : rebase_source : c2075a8657eee795609319b78ad5a7ea3764dfd0