Bug 1893349 - Set windows.notificationserver pref to false to temporarily fix 10 seconds delay for Windows notifications. r=john.bieling
This only removes the 10 seconds delays when clicking on a Windows OS notification.
With these prefs set to `false` the notification is correctly dismissed and the message is visualized.
Unfortunately, the application is not brought to the foreground correctly, but we will take care of this in bug 1863798.

Differential Revision: https://phabricator.services.mozilla.com/D208574

--HG--
extra : amend_source : 441ed8f2688c50caf349c6a064628eb5b4f9f5b1