Bug 1893669 - Finalize support for virtual tag folders and unified mailbox folders. r=aleca
Adding full support for virtual tag folders did force a slight change in
the representation of unified folders. I already tried to not expose
that unified folders actually belong to the "smart" account, but I did
expose the common root folder for all unified folders. After adding the
tags folder, it became obvious, that this should be reverted. The API
should not depend on the implementation detail, that the "tags" folder
is a folder of the same "smart" account.

The unified mailbox folders now have the following representation:
 id: unified://<type>
 path: /unified/<type>

The virtual tag folders have the following representation
 id: tag://<key>
 path: /tag/<key>

This patch adds the missing pieces for full support of virtual tag
folders and the required tests.

Differential Revision: https://phabricator.services.mozilla.com/D208945

--HG--
extra : moz-landing-system : lando