Bug 1896009 - Ignore changes to TreeView header height when the view is hidden. r=mkmelin
When the view gets hidden, the resize observers are informed that its height is 0. Consequently
when it becomes unhidden, some calculations are performed using a cached height of 0, before the
resize observers are updated.

When the header itself is hidden, we DO cache a height of 0, because that is the actual height.

Differential Revision: https://phabricator.services.mozilla.com/D210245

--HG--
extra : amend_source : 28fb039d266f9884f4c3cd768e003f23a28fd6ec