Bug 1898853 - Add function to convert html snippets to plaintext. r=mkmelin
This adds the function to convert HTML snippets to plain text. It was not
added as an option of `messages.listTextParts()`, because it can be used
in other scenarios as well. Introducing it as a stand-alone function
without being bound to a specific message increases its usefulness.

Differential Revision: https://phabricator.services.mozilla.com/D211001

--HG--
extra : moz-landing-system : lando