Bug 1830212 - Add test for a listInlineTextParts() function, which is not part of the messages API, but will be used in documentation. r=aleca
The `listInlineTextParts()` function defined in this test is a helper
function to extract the content of inline text parts. These parts are
the only parts which have their content returned by `messages.getFull()`
and are not returned by `messages.listAttachments()`.

Since developers need detailed knowledge about MIME and the structure of
messages in order to access the content of these parts via `getFull()`,
this helper function can be used to return a flat list of all inline
text parts, just as `messages.listAttachments()` returns a flat list of
attachments.

There is an attempt to add this helper function directly to the API (see
bug 1898854), but it is currently stuck in review. For now, the helper
function will be provided through documentation. Since we are officially
exposing this function, this patch adds a test for it.

Differential Revision: https://phabricator.services.mozilla.com/D212562

--HG--
extra : amend_source : 5af9e60ab359bd823ef000f149021aab2c7e3e07