Bug 1901651 - Additional Rust automation messages to sheriffs. r=dandarnell
The standard Taskcluster notifications have been recently expanded.
"on-any" is deprecated and replaced with "on-resolved" and fires when a job
ends regardless of whether it ran successfully or failed.
"on-defined" is new and fires when a job is defined by the decision task.

Now there's a notification that the automation will run, followed by
a notification mid-way when Rust update is needed.

The final notification that a Phab revision is ready remains.

The "on-defined" notification will still come probably 10 minutes after the
initial m-c push notification, but it's better than the 40 minute delay
currently observed.

Differential Revision: https://phabricator.services.mozilla.com/D215137

--HG--
extra : amend_source : 403d99e26f3d3752a0f2ba1b1085e8e08a8ec63a