Bug 1896171 - Part 6: Send outgoing message to Bcc recipients when using EWS. r=leftmostcat
EWS requires Bcc recipients to be passed separately from the others. This is
because non-Bcc recipients are included in the MIME content, and therefore don't
need to be included separately in the request. But for compliance and privacy,
we don't include Bcc recipients in the MIME content.

I've also used this opportunity to make the type of the recipients a bit more
structured, which isn't a huge flow change since we would end up doing the same
parsing on the SMTP server side. So this is a small change that ends up saving a
bit of complexity on the server implementation's side.

Differential Revision: https://phabricator.services.mozilla.com/D212838

--HG--
extra : moz-landing-system : lando