Bug 1909617 - Don't show "Mark Folder Read" for virtual folders with search terms. r=aleca
Without a view, there seems to be no reasonable way to mark as read only the searched messages in
the underlying folders. So I suggest this approach. The functionality is still provided by the
"Mark All Read" command when the folder is displayed (which should be fixed in bug 1909617).

Differential Revision: https://phabricator.services.mozilla.com/D218486

--HG--
extra : amend_source : d5b0cd8343b7f12e0417f3566f96cca91fbd92e1