Bug 1919265 - Force Unix-style line endings when generating Rust manifests. r=aleca
This change sets the `newline` argument to `open` calls to `\n`, so that
any line ending in these files are Unix-style; as opposed to Python's
default behaviour which is to use the line ending style of the operating
system.

This patch only changes `open` calls that are made in the context of
writing into a file (as opposed to reading from one).

Differential Revision: https://phabricator.services.mozilla.com/D222459

--HG--
extra : amend_source : 7443a9d06f741803c7cede87fbaa321a0040cf96