Bug 1877640 - Fix Updating event doesn't always send notification to attendees r=mkmelin
If the attendee list contains the organizer of the calendar, the event is incorrectly recognized as an unanswered event and the event is updated before it is processed. This sets the dirty flag for the oldItem. When stampTime is checked, the timestamp is set to the same value so that no difference is recognized.

Differential Revision: https://phabricator.services.mozilla.com/D208084

--HG--
extra : moz-landing-system : lando