Bug 1918885 - Fix regressions in the mailTabs and messageDisplay API related to the selection of threads and context-clicks in the thread pane. r=mkmelin
If a collapsed thread is selected, then `mailTabs.getSelectedMessages()`
used to return all messages, including those of the collapsed thread.
This is no longer the case in Thunderbird 128. The original behavior
is restored by this patch.

A related but very old regression fixed by this patch is the behavior of
`messageDisplay.getDisplayedMessages()` when there is a message being
context-clicked. Instead of returning the actual displayed message, it
returned the message currently being context-clicked. The expected
behavior to always return the displayed message is restored by this
patch.

This patch adds a test to ensure the expected behavior.

Differential Revision: https://phabricator.services.mozilla.com/D223739

--HG--
extra : amend_source : 57585100d1418cfb97bed1aff1863d0b1d2b5891