Bug 1918925 - Use x-ray vision only for known nested attachments. r=mkmelin
jsmime uses a special format for the partName of a part of a nested
message: `3$.2`

This is addressing the 2nd attachment of an attached messages which
itself is the 3rd attachment. Thunderbird is prettifying that partName
to `1.3.1.2` in its message URLs. Before passing such a partName to
jsmime, the prettified format was "decoded" again:

```
let partNum = partName.split(".").slice(1).join(".").replaceAll(".1.", "$.")
```

A hierarchy represented by a partName of the format `1.x.1.x...` was
assumed to be always due to nested messages. However, that assumption
was wrong. It is legal to have nested `multipart/*` parts which could
generate such a partName without a nested message being involved.
Forcefully "decoding" such a partName causes jsmime to not find the
requested attachment, resulting in the reported bug.

To fix this, we no longer auto-decode such partNames, but only when we
know we want to access a nested attachment. We also add an additional
parsing step, to see all parts returned by jsmime and match the one we
are interested in.

Nested attachments are not returned by listAttachments(). X-ray vision
however was accessible through the WebExtension API by "guessing" the
partName of a nested attachment. This allowed us to test x-ray vision.
I would like to keep being able to do that, but since we no longer
"auto-decode" nested partNames, the caller has to explicitly request
x-ray vision by specifying a x-ray partName, for example `1.3$.2`. The
schema definition was adjusted to allow such partNames to be used. That
feature will remain undocumented.

Differential Revision: https://phabricator.services.mozilla.com/D223581

--HG--
extra : amend_source : 4050bc3940644664476dbb167da9e1c83aff9996