Bug 1921765 - Use special care for decoding address headers in messages.getFull(). r=mkmelin
This is a follow up to Bug 1919244. For convenience, the API has always
decoded the returned headers in `messages.getFull()`. That turned out to
make it impossible to parse certain headers. For example, the following
TO header
```
To: =?UTF-8?Q?H=C3=B6rst=2C_Kenny?= <K.Hoerst@invalid>, new@thunderbird.bug
```
was decoded to
```
To: Hörst, Kenny <K.Hoerst@invalid>, new@thunderbird.bug
```
This header can no longer be parsed correctly, because of the comma in
the first display name. The display name should now be quoted. The issue
seems to be specific to address headers and there is no global comma
based fix we could apply to all headers. Similar to jsmime [1], which is
using a dedicated parser for well known address headers, we will handle
these address headers separately as well. It is obviously not a perfect
fix, but the best we can do.

[1] : https://searchfox.org/comm-central/rev/e831583b72010c3ccce7d8ce928f2e54e905687d/mailnews/mime/jsmime/jsmime.mjs#154-177

Differential Revision: https://phabricator.services.mozilla.com/D224115

--HG--
extra : moz-landing-system : lando