Bug 1921366 - Implement initial Gloda boilerplate code to get a conversation view. r=darktrojan
Initial very simple boilerplate implementation of conversation view.
Currently leveraging the multiMessageBrowser view to have a quick and
easy UI to plug in. This will later be changed.

### Known issues:

- `Gloda.getMessageCollectionForHeader()` returns duplicated sent messages.
- Sent messages cannot be expanded when clicking on them in the thread view.
- No visual distinction between sent and received messages.
- No visual indication of messages location in other folders.

This is all behind an experimental pref so the objective is to ensure
the initial logic and used conditions are correct as an iterative step.

Differential Revision: https://phabricator.services.mozilla.com/D223885

--HG--
extra : moz-landing-system : lando