Bug 1924869 - Improve readability of MimeJSComponents.sys.mjs. r=freaktechnik
This is not a functional change.

`HEADER_OPTION_DECODE_2231` is only supported if used together with
`HEADER_PARAMETER` (which is correctly stated in mimeParser.sys.mjs). It
should not be specified in places where it is not having an effect to
not trick the reader into thinking decoding according to RFC 2231 is
supported.

Differential Revision: https://phabricator.services.mozilla.com/D225717

--HG--
extra : amend_source : f3dbe79f6eb3639fe286a9f48714139a469b6cd3