Bug 1921658 - Set CalendarBaseView.weekStartOffset after the default preferences. r=aleca
In the earliest case, `ensureInitialized` runs as tabs are restored, so always after
`setLocaleDefaultPreferences`. Unlike `connectedCallback` which runs as soon as the custom element
is defined and attached to the DOM.

It needs to happen after `setLocaleDefaultPreferences` to get the correct value if the correct
value is not 0.

Differential Revision: https://phabricator.services.mozilla.com/D226086

--HG--
extra : amend_source : 493e0709b2f62a375c56ab3e31a69932401bd1c1