Bug 1923153 - Improve imap keyswap handling. r=mkmelin,leftmostcat
This was found while moving a single IMAP message, which gets a fake
messageKey in the new folder and later triggers the `msgKeyChanged`
event: fakeMessageKey -> serverMessageKey.

There will be no secondary swap for serverMessageKey -> fakeMessageKey.
These mirror swaps only occur, if the code almost guessed the correct
messageKeys while moving multiple messages, and the keys assigned
by the server are swapped.

The patch will register a pending mirror event only if it is really
expected. This also removes the no longer valid entry from the
MessageTracker.

We perform a similar action when attachments are removed from messages.
The core message database is handling the modified message without
attachments as a new message with a new messageKey. The WebExtension API
does not, but updates the tracker information with the new messageKey.
The old tracker information was also not removed and could cause issues.
This might be the reason for Bug 1910483.

Differential Revision: https://phabricator.services.mozilla.com/D224927

--HG--
extra : moz-landing-system : lando