Bug 1787104 - Allow messages.import() to directly import into IMAP folders. r=aleca
The missing link to get `messages.import()` on IMAP folders working is:
```
folder.updateFolder(null);
```
This will trigger some IMAP magic to make the newly added message appear
in the folder and to trigger the `MailServices.mfn.msgAdded` listener.

Another issue was that some protocol implementations did not always
fire the `nsIMsgCopyServiceListener.setMessageKey()` callback. Instead
of trying to fix the implementation, we simply look for messages with
the expected headerMessageId, if we did not get notified about the
messageKey in time.

Differential Revision: https://phabricator.services.mozilla.com/D225342

--HG--
extra : moz-landing-system : lando