Bug 520582 - Adjust unread status of NNTP headers when downloading. r=mkmelin
This patch sets the unread status of downloaded headers according to the newsrc file as soon as they are added to the message database. Previously this was only done when the view was initialized, but `nsMsgDBView::AdjustReadFlag` wasn't called on elided messages in collapsed threads. Especially after repairing the folder or getting the next n messages, this could result in an incorrect and fragile view.

Differential Revision: https://phabricator.services.mozilla.com/D225617

--HG--
extra : amend_source : 26f49a07744acbc8d32597068b7e694ee5932405