Bug 1857413 - Fix browser_subscribe.js by waiting for the right thing. r=mkmelin
This `waitForCondition` call is just waiting for the right number of rows. There could be the right
number of rows part-way through updating the tree, if a row gets added before the row to be removed
gets removed, and this seems to be what happens in most if not all cases.

Differential Revision: https://phabricator.services.mozilla.com/D226594

--HG--
extra : amend_source : 5b3430d088ee7c88be98de51a9b9f2a1e0c5fe7a