Bug 1693772 - Attempt to fix timeout of browser_formPickers.js. r=aleca
I have been analyzing the timeout and see it happen mainly when the
datetime picker should show up. The click seems to be ignored. One
option would be to re-trigger the click in these cases, but that did not
work reliably and was causing other issues. Adding a small delay before
issuing the click worked better.

A different timeout was observed when running the test with --verify,
which sometimes even leads to a crash/hang: In the mailbox case, calling
`input.showPicker()`, the picker opened as a blank square and
`BrowserTestUtils.waitForDateTimePickerPanelShown()` either fulfills
very (very!) late, or crashes. This is not fixed by this patch, but was
also not observed on CI.

Adding some log messages to see the location of the timeout, in case the
intermittent is not fully fixed.

Differential Revision: https://phabricator.services.mozilla.com/D227711

--HG--
extra : amend_source : 97f949aa69c53a597e95fb2c689c0775ce3e548a