Bug 1816238 - Attempt to fix browser_browserContext.js. r=thunderbird-reviewers,babolivier
This is based on a comparison between the following tests:
* /mail/base/test/browser/browser_browserContext.js
* /mail/base/test/widgets/browser_formPickers.js

The two tests look similar. While the first one fails often,
the second one seems to run fine. The only difference I could
find is the second one is waiting for focus before clicking the
extension button. This does seem to fix the first test as well.

Differential Revision: https://phabricator.services.mozilla.com/D228296

--HG--
extra : moz-landing-system : lando