Bug 1927871 - Use BuiltInThemes.ensureBuiltInThemes() instead of manually specifying theme IDs and versions in MailGlue.sys.mjs. r=mkmelin
When the built-in themes are updated, their version numbers change and
all checks in the code to verify if the correct versions are installed
have to be adjusted as well. The location in `MailGlue.sys.mjs` was
forgotten when we last changed the themes in Bug 1844244. It was also
overlooked when we tried to fix this regression for the first time in
Bug 1859271.

Instead of calling `AddonManager.maybeInstallBuiltinAddon()` manually
with hard coded version strings, we now use a dedicated function from
the `BuiltInThemes` module. Thanks Rob Wu for pointing me to it. This
removes the need to update *this* location after future theme changes.

We still have to remember to update the `BuiltInThemes` module after a
theme change. This patch is therefore adding a test to make sure the
correct version strings are used.

Differential Revision: https://phabricator.services.mozilla.com/D228313

--HG--
extra : rebase_source : 78c3c64032305ddb3d1962fc0014e6c822a3e9a1
extra : amend_source : f60ab63fab300ef3199db34d634ba586ef49eaa1