Bug 1816238 - Use readyState to wait for content loads in tests. r=mkmelin
While investigating timeouts of `browser_browserContext.js`, it was
observed that the condition to wait for a load event was triggered by
`browser.webProgress.isLoadingDocument` being `true`, but the load event
was never seen.

The `readyState` however was already `complete`. This patch therefore
changes the wait-for-load condition of content loads, here and in
similar tests.

Differential Revision: https://phabricator.services.mozilla.com/D228429

--HG--
extra : moz-landing-system : lando