Bug 1921974 - Set nodefaultsrc to prevent secondary load of about:blank. r=mkmelin
The reason for the observed missing error page is a race condition
between the load of `about:blank` and the load of the error page. Due to
changes in bug 1911977, the network activity in a browser with a failed
protocol handler load is no longer stopped (to not trigger a network
error page), which in this case does not stop the load of `about:blank`.

The failed WebExtension protocol handler will trigger an error page
nonetheless. The ongoing load of `about:blank` can race against it and
if the error page is loaded faster than `about:blank`, we end up with a
blank page instead of the error page.

More details can be found in the bug.

Differential Revision: https://phabricator.services.mozilla.com/D229469

--HG--
extra : amend_source : 2a543948e3d4c82d5f2179c608ef8ab1909c6be5