Bug 1934979 - Support protocols that don't use nsIMsgMailnewsUrl, such as EWS, when composing from a draft. r=#thunderbird-back-end-reviewers
It turns out it doesn't matter if the URL is an `nsIMsgMailnewsUrl` when the
protocol we're getting the message from doesn't necessarily expect it to be. EWS
doesn't use `nsIMsgMailnewsUrl`-based URLs, and doesn't expect its URLs to
implement it.

This patch also implements `nsIMsgMessageService::StreamMessage` using the util
method added in https://phabricator.services.mozilla.com/D228378.

Differential Revision: https://phabricator.services.mozilla.com/D231284