Bug 1927480 - Fix clang-format on comm-central. Add task to run on whole tree. r=dandarnell
Removing the optimization config on comm-central prior to setting the paths
to format in the command resulted in an invalid command that always succeeded.

Add a variant that can be run through add-new-jobs to run format on the
entire tree for situations like this where mass-update is needed.

Differential Revision: https://phabricator.services.mozilla.com/D228489

--HG--
extra : moz-landing-system : lando