Bug 1936296 - Only log EWS request/response bodies when requested. r=leftmostcat
This was initially included in https://phabricator.services.mozilla.com/D231690,
where it got a passing functional review, but was then split out because there
was a disagreement over whether this was a good idea. See that revision for the
previous discussion.

Differential Revision: https://phabricator.services.mozilla.com/D231805

--HG--
extra : amend_source : 6a06822ff76780ac63998e29272b2f3d1439717f