Bug 1930707 - Remove uncertainty when working with recent folders. r=arschmitz
The term "recent" is not well defined and Thunderbird uses it in two
different places with different meaning:
* "move to / recent" context menu of the thread pane: recently modified
* "recent folders" mode: recently used

From a user perspective, it might be confusing if the folders listed in
the "recent folder" mode and the folders available in the "move to /
recent" context menu are not the same.

This issue became apparent after the WebExtension folders API started to
expose the recently used folders as "recent" as well, and developers
noticed the different return value compared to the "move to / recent"
context menu.

Differential Revision: https://phabricator.services.mozilla.com/D228695

--HG--
extra : amend_source : 4a76dcaa2a1823f7d312c5b139d7af1d9450ebcb