Bug 1913580 - Part 1: Make second parameter of spaces.create() and spaces.update() accept a SpaceTabsProperties object. r=mkmelin
The second parameter of `spaces.create()` and `spaces.update()` is used
to define or update the URL of the default space tab. In order to also
define or update additional tab properties, we now allow to specify a
`SpaceTabsProperties` object.

This patch does not yet add any new functionality, only the alternative
method to specify the URL (as SpaceTabsProperties.url) is added and
tested.

We do have to work around a limitation of the WebExtension framework:
The 2nd and 3rd parameter of `spaces.update()` were optional, so either
the URL or the button properties could be updated. The schema parser was
able to identify which parameter is specified due to their different
types. Now both parameters can be objects and parameter detection is
failing, because the parser is not checking the provided *properties*.

The only way out is to make the 2nd parameter no longer optional and
make it also accept `SpaceButtonProperties` (what the 3rd parameter is
about). There are no negative consequences besides a confusing schema
file.

Differential Revision: https://phabricator.services.mozilla.com/D229815

--HG--
extra : moz-landing-system : lando