Bug 1913580 - Part 2: Update SpaceTracker to store the full SpaceTabProperties object instead of just a URL. r=mkmelin
Part 1 added the possibility to specify a `SpaceTabProperties` object
instead of a URL, but only the URL was actually stored. This patch
updates the SpaceTracker to store the full object.

This patch does not yet add any new functionality.

We do adjust the deprecated spacesToolbar API to no longer store the URL
as part of its SpaceButtonProperties but only in the new
SpaceTabProperties. This might have caused issues in the past, when an
add-on used the spaces API and (!) the spacesToolbar API to
create/update the same space.

Differential Revision: https://phabricator.services.mozilla.com/D229880

--HG--
extra : moz-landing-system : lando