Bug 1802300 - Show net error page if request of POP3 URL fails. r=mkmelin
When downloading the rest of a partial POP3 message fails, display a network error page instead of an empty message pane. This can happen if offline mode is active, or if the UIDL for the message cannot be found on the server.

Differential Revision: https://phabricator.services.mozilla.com/D232646

--HG--
extra : amend_source : a959689cd97220d89e3b58c9eab78ec54bb0d6be