Bug 1921394 - Create a filter for messages with or without a tag. r=mkmelin
This adds functions to SQLite for detecting a string in a space-separated list, and a LiveView filter
for selecting messages with or without a given tag. We'll only use the positive case for now but the
negative case will be useful later.

Differential Revision: https://phabricator.services.mozilla.com/D232180

--HG--
extra : amend_source : 8ea540be090d39624172a1bf15c7f67acacff529